refactor: move utilities to /runtime export
