--- env.d.ts
@@ -1,1 +1,0 @@
-/// <reference path="./playground/.nitro/types/nitro.d.ts" />
